Rename Application Interfaces folder to Contracts
