+++ src/Application/Contracts/IEmailApiService.cs
@@ -1,6 +1,6 @@
 using Merrsoft.MerrMail.Domain.Models;
 
-namespace Merrsoft.MerrMail.Application.Interfaces;
+namespace Merrsoft.MerrMail.Application.Contracts;
 
 public interface IEmailApiService
 {
